@@ -3,7 +3,7 @@ package com.example.emoticon.activity;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import com.example.common.base.BaseActivity;
 import com.example.emoticon.MainActivity;
@@ -14,8 +14,6 @@ public class SplashActivity extends BaseActivity {
         super.onCreate(savedInstanceState);
         //隐藏状态栏
         //getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN);
-
-
         SharedPreferences sharedPreferences = getSharedPreferences("guide", MODE_PRIVATE);
         boolean first = sharedPreferences.getBoolean("firsopen", true);
         if (first) {
