@@ -1,7 +1,7 @@
 package com.example.emoticon.adapter;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -10,30 +10,31 @@ import android.widget.TextView;
 
 import com.example.emoticon.R;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class LabelAdapter extends RecyclerView.Adapter<LabelAdapter.ViewHolder> {
-    List<String> list = new ArrayList<>();
-
+    private List<String> list;
     private OnItemClickListener onItemClickListener;
-    public LabelAdapter(List<String> list){
+
+    public LabelAdapter(List<String> list) {
         this.list = list;
     }
+
     @NonNull
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
-        View view = LayoutInflater.from(viewGroup.getContext()).inflate(R.layout.label_item,viewGroup,false);
+        View view = LayoutInflater.from(viewGroup.getContext()).inflate(R.layout.label_item, viewGroup, false);
         return new ViewHolder(view);
     }
 
     @Override
-    public void onBindViewHolder(@NonNull ViewHolder viewHolder, final int i) {
+    public void onBindViewHolder(@NonNull final ViewHolder viewHolder, int i) {
         viewHolder.textView.setText(list.get(i));
+        viewHolder.itemView.setTag(i);
         viewHolder.delete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                onItemClickListener.OnItemClick(v, i);
+                onItemClickListener.OnItemClick(v, (int) viewHolder.itemView.getTag());
             }
         });
     }
@@ -46,6 +47,7 @@ public class LabelAdapter extends RecyclerView.Adapter<LabelAdapter.ViewHolder> 
     public class ViewHolder extends RecyclerView.ViewHolder {
         TextView textView;
         ImageView delete;
+
         public ViewHolder(@NonNull View itemView) {
             super(itemView);
             textView = itemView.findViewById(R.id.title);
@@ -54,10 +56,11 @@ public class LabelAdapter extends RecyclerView.Adapter<LabelAdapter.ViewHolder> 
     }
 
 
-    public interface OnItemClickListener{
-        void OnItemClick(View view,int position);
+    public interface OnItemClickListener {
+        void OnItemClick(View view, int position);
     }
-    public void setOnItemClickListener(OnItemClickListener onItemClickListener){
+
+    public void setOnItemClickListener(OnItemClickListener onItemClickListener) {
         this.onItemClickListener = onItemClickListener;
     }
 }
