@@ -1,8 +1,8 @@
 package com.example.common.utils;
 
 import android.graphics.drawable.Drawable;
-import android.support.annotation.NonNull;
-import android.support.v4.graphics.drawable.DrawableCompat;
+import androidx.annotation.NonNull;
+import androidx.core.graphics.drawable.DrawableCompat;
 
 public class DrawableTintUtil {
     public static Drawable tintDrawable(@NonNull Drawable drawable,int color){
