@@ -1,8 +1,9 @@
 package com.example.emoticon.activity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
 import android.view.View;
+
+import androidx.annotation.Nullable;
 
 import com.example.common.base.BaseActivity;
 import com.example.common.manager.UserManager;
