@@ -1,7 +1,7 @@
 package com.example.common.widget;
 
 import android.content.Context;
-import android.support.design.widget.TabLayout;
+import com.google.android.material.tabs.TabLayout;
 import android.util.AttributeSet;
 
 /**
