@@ -1,5 +1,7 @@
 package com.example.emotion.user.contract;
 
+import androidx.annotation.NonNull;
+
 import com.example.common.base.BasePresenter;
 import com.example.common.base.BaseView;
 import com.example.common.bean.Emoticon;
@@ -14,11 +16,25 @@ import java.util.List;
  */
 public interface UserEmoticonsContract {
     interface EmoticonView extends BaseView<UserEmoticonsContract.Presenter>{
-        void onError(Exception e);
-        void onFinish(List<Emoticon> emoticonList);
+        /**
+         * 获取失败回调
+         * @param e 错误内容
+         */
+        void onError(@NonNull Exception e);
+
+        /**
+         * 获取成功回调
+         * @param emoticonList 获取的内容
+         */
+        void onFinish(@NonNull List<Emoticon> emoticonList);
     }
 
-    interface Presenter extends BasePresenter{
+    interface Presenter extends BasePresenter<Emoticon>{
+        /**
+         * 获取数据
+         * @param userId 用户ID
+         * @param skip 跳过几个数据
+         */
         void loadData(int userId, int skip);
     }
 }
