@@ -18,13 +18,18 @@ import retrofit2.Call;
  * PS:
  */
 public class UserEmoticonsPresenter implements UserEmoticonsContract.Presenter {
-    UserEmoticonsContract.EmoticonView emoticonView;
+    private UserEmoticonsContract.EmoticonView emoticonView;
 
     public UserEmoticonsPresenter(UserEmoticonsContract.EmoticonView emoticonView) {
         this.emoticonView = emoticonView;
         emoticonView.setPresenter(this);
     }
 
+    /**
+     * 获取用户表情数据
+     * @param userId 用户ID
+     * @param skip 跳过几个数据
+     */
     @Override
     public void loadData(int userId, int skip) {
         UserProtocol userProtocol = RetroClient.getServices(UserProtocol.class);
